requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
datadog==0.29.3
decorator==4.4.0
idna==2.8
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py-trello==0.15.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
